@@ -1,7 +1,7 @@
 source 'http://rubygems.org'
 ruby "2.3.1"
 
-gem 'rails', '4.0.4'
+gem 'rails', '~> 4.1.0'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
@@ -74,8 +74,8 @@ gem 'simple_form', '~> 3.1.0.rc1'
 gem 'country_select'
 gem 'wicked', '~> 1.2'
 gem 'wicked_pdf' # for PDF generation
-gem 'activeadmin', github: 'gregbell/active_admin'
-gem 'ransack', '~> 1.1.0'
+gem 'activeadmin', '~> 1.0.0.pre4'
+gem 'ransack'
 
 gem 'comfortable_mexican_sofa'
 gem "newrelic_rpm"
@@ -119,6 +119,10 @@ group :test, :development do
   gem 'pry-byebug'
   gem 'better_errors'
   gem "binding_of_caller"
+end
+
+group :development do
+  gem 'web-console', '~> 2.0'
 end
 
 group :test do
